Compile shared library when precompiling a package
Add natived flag to the code instance and teach dump to serialize and
deserialize this flag and native function names when the flag is set.
Compile the shadow_output to a native shared library when
precompiling a package.

testing precompiling with function names

globals

rebased on codegen patch, but broken

changes